requirements.txt
playwright>=1.47
python-dotenv>=1.0
pillow>=10.3
pyyaml>=6.0.1
